Conflicts:
	Classes/ConfigFiles/SEBSettings.m
	SEB/AppDelegate.m
	SEB/Base.lproj/Main.storyboard
	SEB/Classes/AdditionalViews/AboutSEBiOSViewController.m
	SEB/Classes/AdditionalViews/SEBUIController.m
	SEB/Classes/BrowserComponents/SEBBrowserTabViewController.h
	SEB/Classes/BrowserComponents/SEBBrowserTabViewController.m
	SEB/Classes/BrowserComponents/SEBWebViewController.m
	SEB/Resources/SEBDev-Info.plist
	SEB/SEBViewController.h
	SEB/SEBViewController.m
	SEBTests/Info.plist
	SEBUITests/Info.plist
	SafeExamBrowser-Info.plist
	SafeExamBrowser.xcodeproj/project.pbxproj

+++ Classes/Cryptography/MscX509Common/MscRSAKey.m
@@ -55,7 +55,7 @@
                 @throw [MscX509CommonLocalException exceptionWithCode:FailedToGenerateKey];
             }
         
-            returnCode = RSA_generate_key_ex(rsa, keySize, bigNumber, NULL);
+            returnCode = RSA_generate_key_ex(rsa, (int)keySize, bigNumber, NULL);
             if (returnCode != 1) {
                 NSLog(@"Failed to generate RSA key, function RSA_generate_key_ex returned with %d", returnCode);
                 @throw [MscX509CommonLocalException exceptionWithCode:FailedToGenerateKey];
